requirements.txt
Flask==2.3.3
celery==5.3.4
redis==5.0.1
flask-socketio==5.3.6
openai==1.3.7
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
PyPDF2==3.0.1
Pillow==10.0.1
python-dotenv==1.0.0
flask-cors==4.0.0
plotly==5.17.0
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3